NICO_AI_PHASE_1/0_text_extraction_scripts/requirements.txt
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
indic-nlp-library>=0.92
